requirements.txt
crewai[anthropic] 
streamlit 
requests 
python-dotenv 
